core (back): extract event but using a switch case
