+++ back/game/event/event.go
@@ -1,7 +1,6 @@
-package game
+package event
 
 // Event A game event
 type Event interface {
 	EventMessage() string
-	Apply(game Game) Game
 }
