+++ back/game/game.go
@@ -1,9 +1,13 @@
 package game
 
+import (
+	"github.com/ThomasFerro/armadora/game/event"
+)
+
 // Game Instance of an Aramdora game
 type Game interface {
 	State() State
-	Apply(event GameCreated) Game
+	Apply(event event.GameCreated) Game
 }
 
 type game struct {
@@ -15,22 +19,26 @@ func (g game) State() State {
 	return g.state
 }
 
-func (g game) Apply(event GameCreated) Game {
+func (g game) Apply(event event.GameCreated) Game {
 	g.state = WaitingForPlayers
 	return g
 }
 
 // CreateGame Create a new game
-func CreateGame() GameCreated {
-	return GameCreated{}
+func CreateGame() event.GameCreated {
+	return event.GameCreated{}
 }
 
 // ReplayHistory Replay the provided history to retrieve the game state
-func ReplayHistory(history []Event) Game {
+func ReplayHistory(history []event.Event) Game {
 	var returnedGame Game
 	returnedGame = game{}
-	for _, event := range history {
-		returnedGame = event.Apply(returnedGame)
+	for _, nextEvent := range history {
+		switch nextEvent.(type) {
+		case event.GameCreated:
+			gameCreatedEvent, _ := nextEvent.(event.GameCreated)
+			returnedGame = returnedGame.Apply(gameCreatedEvent)
+		}
 	}
 	return returnedGame
 }
